Added restaurant title to db; displaying title
